requirements.txt
arrow==1.3.0
binaryornot==0.4.4
briefcase==0.3.18
build==1.2.1
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
clr-loader==0.2.6
colorama==0.4.6
cookiecutter==2.6.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
packaging==24.0
platformdirs==4.2.2
pluggy==1.5.0
psutil==5.9.8
pycparser==2.22
Pygments==2.18.0
pyproject_hooks==1.1.0
pytest==8.2.2
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pythonnet==3.0.3
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
six==1.16.0
smmap==5.0.1
text-unidecode==1.3
toga-core==0.4.4
toga-winforms==0.4.4
tomli_w==1.0.0
travertino==0.3.0
types-python-dateutil==2.9.0.20240316
urllib3==2.2.1
